SCALE-SEAM (104) — coverage maps, client.ts 572 -> 554
spineTraceability + scopeRegister -> coverageMaps.ts: "how completely is this
project's chain of records linked, and exactly what is missing".

THIS IS THE AFFIRMATIVE HALF OF A BOUNDARY (103) DREW BY EXCLUSION.
acceptanceGates.ts took the four methods that collapse the project to one
accept/refuse verdict, and named spineTraceability as the closest miss PRECISELY
because it maps completeness rather than deciding. Both members here report
proportion-complete across several link types (specs_packaged_pct /
packages_costed_pct / sheets_specced_pct / spec_to_budget_pct; pct_quantified /
pct_allocated / pct_scheduled) plus the individual records lacking the link
(specs_without_bid_package / bid_packages_without_cost_code / sheets_without_spec;
gap_items with per-item gaps[] and status complete|gap). Neither carries a
verdict field. Same test as (103), run for inclusion this time.

DERIVING THE POPULATION REQUIRED RESOLVING NAMED RETURN TYPES, AND THAT IS THE
LESSON. A scan of method bodies for coverage vocabulary returned scopeRegister,
citedQuery and progressActuals -- and MISSED spineTraceability, the strongest
member, because its return is the named type SpineTraceability and its body
contains none of those words. That is the exact MIRROR of the mistake (103)
recorded, where a body scan OVER-counted by matching the doc comment of the next
method. A textual scan of a typed language reads neither the comments nor the
types correctly: it finds candidates, it never counts them.

EXCLUSIONS. citedQuery carries coverage / uncited_claims / fully_cited, but its
product is an ANSWER whose sourcing those annotate -- the caller renders a cited
answer, not a worklist. masterBuilderBrief is THE CLOSEST CALL in this slice,
closer than spineTraceability was to the gates, since it has readiness_pct,
ready_steps and gap_steps; it stays because its product is a narrative brief with
a reframe_prompt and a disclaimer, where the percentages head a document rather
than being it -- recorded rather than asserted, since a later reader may decide
otherwise. progressActuals pct_complete is physical progress against a planned
quantity banded ahead/on_track/behind: variance over time, not completeness of
linkage.

A FOLLOW-UP PROPOSED IN (103) IS WITHDRAWN, AND THAT IS A CORRECTION. #443 said
three consecutive slices leaving an orphaned type import was "a pattern that
wants a gate rather than vigilance". It already has one: tsc reports the orphan
as TS6196 and caught every instance -- EnergyResult (101), PrequalScores (102),
DiligenceReadiness (103) and now SpineTraceability (104). None reached a commit.
Proposing a gate for a class the toolchain already gates is how a checklist grows
without getting safer.

Verified: 2072 web tests / 205 files, tsc + eslint clean, whole-tree ruff clean,
ratchet mutation-checked at 553 (3 assertions fire), 49-entry narrative chain
unbroken, and test_claude_md_gates + test_route_reachability + test_reachable +
test_import_cycles + test_doc_substance + test_ruff_scope + test_declared_imports
all green.
